fix: Clean up code, add javadoc documentation, rename packages, add validation constraints, move common logic to common-dtos module
FEA-17 Completed

+++ backend/common-dtos/src/main/java/cz/cvut/fel/dto/OrderItemDto.java
@@ -7,13 +7,18 @@ import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
 
+/**
+ * Data Transfer Object representing an item in an order.
+ *
+ * Contains product details, including product ID, product name, quantity, price, and the order ID it belongs to.
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class OrderItemDto {
     private String productId;
-    private Integer orderId;
+    private Long orderId;
     private String productName;
     private BigDecimal quantity;
     private BigDecimal price;
